Fix Clippy lints
